remove standard interfaces from the beta branch
