+++ src/index.html
@@ -28,8 +28,6 @@
             <option value="cluster3" title="Cluster Pro">Cluster Pro</option>
             <option value="pulsar" title="Pulsar">Pulsar</option>
             <option value="transient" title="Transient">Transient</option>
-            <option value="pulsarTest" title="PulsarTest">PulsarTest</option>
-            <option value="variableTest" title="VariableTest">VariableTest</option>
           </select>
         </form>
         <div id="input-div"></div>
